refactor(tests,#10066): consolidate test_extract_pptx_titles — delete legacy shadow, port 2 unique tests
The legacy scripts/tests/test_extract_pptx_titles.py shadowed the canonical
scripts/notebook_tools/tests/ copy via a Python module-name collision
(sys.modules cached the legacy module first), silently suppressing the
canonical file's 5 TestSlideMarker regex unit tests + 5 canon-only TestExtract
variants (10 tests dead whenever both ran). Ported the 2 legacy-unique tests
(test_truncation_at_80, test_first_h1_wins) into the canonical TestExtract
class, then deleted the shadow. Net coverage 10 -> 15 running tests.

Per-test analyze->merge->archive (CLAUDE.md global 'Consolider'). Count
(canon 13 > legacy 10) was not proof of strict superset — 2 unique tests
existed despite the count. Tranche 1 of #10066 (12 duplicated files).

Validation: pytest canon alone -> 15 passed; notebook_tools/tests/ subset
-> 4210 passed, 2 skipped; no dangling refs to deleted path.
